requirements.txt
anyio==4.0.0
astropy==5.3.4
astroquery==0.4.6
attrs==23.1.0
beautifulsoup4==4.12.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
contourpy==1.1.1
cryptography==41.0.4
cycler==0.12.0
exceptiongroup==1.1.3
fonttools==4.43.0
h11==0.14.0
html5lib==1.1
httpcore==0.18.0
httpx==0.25.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.0
jaraco.classes==3.3.0
jeepney==0.8.0
jplephem==2.19
keyring==24.2.0
kiwisolver==1.4.5
linear-series==7
llvmlite==0.41.0
Logbook==1.7.0
lxml==4.9.3
matplotlib==3.7.3
more-itertools==10.1.0
numba==0.58.0
numpy==1.25.2
orbital-surface==2
packaging==23.2
pandas==2.1.1
Pillow==10.0.1
plotly==5.17.0
poliastro==0.17.0
pycparser==2.21
pyerfa==2.0.0.3
pyorbital==1.8.0
pyparsing==3.1.1
pyproj==3.6.1
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.2
pyshp==2.3.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytils==0.4.1
pytz==2023.3.post1
pyvo==1.4.2
PyYAML==6.0.1
Represent==1.6.0.post0
requests==2.31.0
rush==2021.4.0
scipy==1.11.3
SecretStorage==3.3.3
sgp4==2.22
six==1.16.0
skyfield==1.46
sniffio==1.3.0
soupsieve==2.5
spacetrack==1.0.1
tenacity==8.2.3
TLE-tools==0.2.3
tools==0.1.9
tzdata==2023.3
urllib3==2.0.6
webencodings==0.5.1
zipp==3.17.0
